requirements.txt
pandas
numpy
opencv-python
scikit-learn-extra
scikit-image
Pillow
matplotlib